requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_socketio
flask_cors
psycopg2-binary
python-dotenv
greenlet
requests
eventlet